refactor : `app`모듈 `namespace` `team.jsv.app` → `team.jsv.icec`로 변경 (#33)
