+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
     <uses-feature android:name="android.hardware.camera.any" />
 
     <application
-        android:name=".IcecApplication"
+        android:name="team.jsv.icec.IcecApplication"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
